requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
